gorg: several things
1. fix a major bug in the call to polyhedron: need to list the
faces in clockwise order as viewed from outside the 3d surface
in order for build to succed. Otherwise you get this error:

"WARNING: Object may not be a valid 2-manifold and may need repair!"

2. refactor so that I include/use the common modules but render
the top and bottom of the Super-Chip in separate files so that I
can 3d print them separately and assemble.